backend/requirements.txt
fastapi
uvicorn
pyairtable
python-dotenv
pydantic
httpx
httpie
pyjwt
python-jose
passlib
